LMS-1988 Make file commands more robust.
SVN: 19778

+++ datastore_server/source/java/ch/systemsx/cisd/etlserver/registrator/api/v1/impl/MkdirsCommand.java
@@ -50,6 +50,12 @@ class MkdirsCommand extends AbstractTransactionalCommand
     {
         File src = new File(directoryAbsoutePath);
 
+        if (false == src.exists())
+        {
+            // Nothing to rollback
+            return;
+        }
+
         IFileOperations fileOperations = FileOperations.getMonitoredInstanceForCurrentThread();
         fileOperations.delete(src);
     }
